requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4

# AI/ML Dependencies
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1

# Computer Vision
ultralytics==8.0.206
mediapipe==0.10.7
face-recognition==1.3.0

# Audio Processing
pyaudio==0.2.11
webrtcvad==2.0.10

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
websockets==12.0
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
